chore(plot): move LegendManager into bokeh_source.py
