+++ src/caret_analyze/plot/visualize_lib/bokeh/bokeh.py
@@ -14,23 +14,22 @@
 
 from __future__ import annotations
 
-from collections import defaultdict
 from logging import getLogger
 from typing import Any, Dict, Generator, List, Optional, Sequence, Tuple, Union
 
 from bokeh.colors import Color, RGB
-from bokeh.models import (GlyphRenderer, HoverTool, Legend,
-                          LinearAxis, Range1d, SingleIntervalTicker)
+from bokeh.models import HoverTool, LinearAxis, Range1d, SingleIntervalTicker
 from bokeh.plotting import ColumnDataSource, Figure, figure
 
 import colorcet as cc
 import numpy as np
 
-from .visualize_lib_interface import VisualizeLibInterface
-from ..metrics_base import MetricsBase
-from ...record import RecordsInterface
-from ...runtime import (CallbackBase, Communication, Publisher,
-                        Subscription, SubscriptionCallback, TimerCallback)
+from .bokeh_source import LegendManager
+from ..visualize_lib_interface import VisualizeLibInterface
+from ...metrics_base import MetricsBase
+from ....record import RecordsInterface
+from ....runtime import (CallbackBase, Communication, Publisher, Subscription,
+                         SubscriptionCallback, TimerCallback)
 
 TimeSeriesTypes = Union[CallbackBase, Communication, Union[Publisher, Subscription]]
 
@@ -369,74 +368,3 @@ class LineSource:
                 data_dict[k] = [self._get_description(k, target_object)]
 
         return data_dict
-
-
-class LegendManager:
-    """Class that manages legend in Bokeh figure."""
-
-    def __init__(self) -> None:
-        self._legend_count_map: Dict[str, int] = defaultdict(int)
-        self._legend_items: List[Tuple[str, List[GlyphRenderer]]] = []
-        self._legend: Dict[Any, str] = {}
-
-    def add_legend(
-        self,
-        target_object: Any,
-        renderer: GlyphRenderer
-    ) -> None:
-        """
-        Store a legend of the input object internally.
-
-        Parameters
-        ----------
-        target_object : Any
-            Instance of any class.
-        renderer : bokeh.models.GlyphRenderer
-            Instance of renderer.
-
-        """
-        label = self.get_label(target_object)
-        self._legend_items.append((label, [renderer]))
-        self._legend[target_object] = label
-
-    def draw_legends(
-        self,
-        figure: Figure,
-        max_legends: int = 20,
-        full_legends: bool = False
-    ) -> None:
-        """
-        Add legends to the input figure.
-
-        Parameters
-        ----------
-        figure : Figure
-            Target figure.
-        max_legends : int, optional
-            Maximum number of legends to display, by default 20.
-        full_legends : bool, optional
-            Display all legends even if they exceed max_legends, by default False.
-
-        """
-        for i in range(0, len(self._legend_items)+10, 10):
-            if not full_legends and i >= max_legends:
-                logger.warning(
-                    f'The maximum number of legends drawn by default is {max_legends}. '
-                    'If you want all legends to be displayed, '
-                    'please specify the `full_legends` option to True.'
-                )
-                break
-            figure.add_layout(Legend(items=self._legend_items[i:i+10]), 'right')
-
-        figure.legend.click_policy = 'hide'
-
-    def get_label(self, target_object: Any) -> str:
-        if target_object in self._legend:
-            return self._legend[target_object]
-
-        class_name = type(target_object).__name__
-        label = f'{class_name.lower()}{self._legend_count_map[class_name]}'
-        self._legend_count_map[class_name] += 1
-        self._legend[target_object] = label
-
-        return label
